requirements.txt
# EFMD Gap Analysis Tool - Requirements
# =====================================

# Core
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=1.0.0

# API
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6

# Frontend
streamlit>=1.25.0

# AI Services
anthropic>=0.18.0
google-generativeai>=0.3.0

# Database
supabase>=2.0.0

# Document Processing
pypdf>=3.0.0
python-docx>=0.8.11

# PDF Report Generation (optional)
weasyprint>=60.0

# Development
pytest>=7.0.0
